+++ Tests/Unit/MSFT_xMsiPackage.Tests.ps1
@@ -282,6 +282,26 @@ Describe 'xMsiPackage Unit Tests' {
                                              -ShouldThrow $false
             }
 
+            Context 'Reboot handling' {
+                Mock -CommandName 'Start-MsiProcess' -MockWith { return 3010 }
+                Mock -CommandName 'Set-DSCMachineRebootRequired' -MockWith {}
+
+                It 'Should request reboot by default' {
+                    $setTargetResourceParameters.IgnoreReboot = $false
+                    { $null = Set-TargetResource @setTargetResourceParameters } | Should -Not -Throw
+
+                    Assert-MockCalled -CommandName 'Set-DSCMachineRebootRequired' -Exactly 1 -Scope 'It'
+                }
+
+                It 'Should not request reboot if IgnoreReboot specified' {
+                    $setTargetResourceParameters.IgnoreReboot = $true
+                    { $null = Set-TargetResource @setTargetResourceParameters } | Should -Not -Throw
+
+                    Assert-MockCalled -CommandName 'Set-DSCMachineRebootRequired' -Exactly 0 -Scope 'It'
+                }
+
+            }
+
             $setTargetResourceParameters.Ensure = 'Absent'
 
             # The URI scheme doesn't matter for uninstallation - it will always do the same thing